backend/server.py
import json
import os
from datetime import datetime
from doctest import debug
from pathlib import Path
from typing import Any, Dict, List, Optional

import uvicorn
from dotenv import load_dotenv
from fastapi import Body, Depends, FastAPI, File, Form, HTTPException, UploadFile
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
from pydantic import BaseModel

# 导入AI服务
from services.ai_service import AIService

# 创建FastAPI应用
app = FastAPI(title="AI小说编辑器API")

# 配置CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # 在生产环境中应该设置为特定域名
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 创建数据目录
os.makedirs("data/stories", exist_ok=True)

# 初始化AI服务
ai_service = AIService()


# 定义数据模型
class Story(BaseModel):
    id: str
    title: str
    content: str

class AIRequest(BaseModel):
    story_content: str
    request_type: str
    specific_request: Optional[str] = None

# 路由定义

@app.get("/")
async def read_root():
    return {"message": "Welcome to the AI小说编辑器API"}

# 获取所有小说
@app.get("/stories")
async def get_stories():
    # 实现获取所有小说的逻辑
    pass

# 获取单个小说
@app.get("/stories/{story_id}")
async def get_story(story_id: str):
    # 实现获取单个小说的逻辑
    pass

# 创建新小说
@app.post("/stories")
async def create_story(story: Story):
    # 实现创建新小说的逻辑
    pass

# 更新小说
@app.put("/stories/{story_id}")
async def update_story(story_id: str, story: Story):
    # 实现更新小说的逻辑
    pass

# 删除小说
@app.delete("/stories/{story_id}")
async def delete_story(story_id: str):
    # 实现删除小说的逻辑
    pass

# AI辅助功能
@app.post("/ai/assist")
async def ai_assist(request: AIRequest):
    # 实现AI辅助功能的逻辑
    if request.request_type == 'plot':
        # 情节建议逻辑
        pass
    elif request.request_type == 'character':
        # 角色发展逻辑
        pass
    elif request.request_type == 'style':
        # 写作风格逻辑
        pass
    elif request.request_type == 'body':
        # 生成正文逻辑
        pass
    else:
        raise HTTPException(status_code=400, detail="无效的请求类型")

# 新增：生成标题
@app.post("/ai/generate-title")
async def generate_title(tags: List[str] = Body(...)):
    # 实现生成标题的逻辑
    pass

# 新增：生成大纲
@app.post("/ai/generate-outline")
async def generate_outline(title: str = Body(...), tags: List[str] = Body(...)):
    # 实现生成大纲的逻辑
    pass

# 新增：生成章节梗概
@app.post("/ai/generate-chapter-summaries")
async def generate_chapter_summaries(outline: str = Body(...)):
    # 实现生成章节梗概的逻辑
    pass

if __name__ == "__main__":
    uvicorn.run("server:app", host="0.0.0.0", port=8000, reload=True)