pyproject.toml
[project]
name = "ai-novel-editor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "openai==1.76",
    "pydantic==2.4.2",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "uvicorn==0.23.2",
]
